refactor: organize chat package into api application and domain
