+++ src/main/java/com/codenames/codenames/backend/chat/api/ChatController.java
@@ -1,6 +1,8 @@
-package com.codenames.codenames.backend.chat;
+package com.codenames.codenames.backend.chat.api;
 
+import com.codenames.codenames.backend.chat.api.dto.ChatDto;
 import com.codenames.codenames.backend.lobby.application.LobbyService;
+import com.codenames.codenames.backend.chat.application.ChatService;
 import com.codenames.codenames.backend.chat.api.dto.ChatMessageType;
 import com.codenames.codenames.backend.lobby.domain.Role;
 import com.codenames.codenames.backend.lobby.domain.Team;
