+++ src/test/java/com/codenames/codenames/backend/chat/domain/ChatHistoryTest.java
@@ -1,8 +1,9 @@
-package com.codenames.codenames.backend.chat;
+package com.codenames.codenames.backend.chat.domain;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
+import com.codenames.codenames.backend.chat.api.dto.ChatDto;
 import com.codenames.codenames.backend.chat.api.dto.ChatMessageType;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
