+++ src/test/java/com/codenames/codenames/backend/lobby/application/LobbyServiceTest.java
@@ -12,7 +12,7 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import com.codenames.codenames.backend.chat.ChatService;
+import com.codenames.codenames.backend.chat.application.ChatService;
 import com.codenames.codenames.backend.game.application.GameService;
 import com.codenames.codenames.backend.lobby.domain.Lobby;
 import com.codenames.codenames.backend.lobby.domain.Player;
